+++ src/lib.rs
@@ -91,7 +91,6 @@ pub use nodemap::{
     Attributes, Entry, MutableAttributes, MutableNamespaces, MutableNodeMap, Namespaces, NodeMap,
 };
 pub use parse::{Span, SpanInfo, SpanInfoKey};
-pub use serialize::PrettyOutputToken;
 pub use xmlvalue::{
     Attribute, Comment, Element, Namespace, Prefixes, ProcessingInstruction, Text, Value, ValueType,
 };